wip(feat!): introduce frame iterator
Prior to this change, the crate would attempt to compute the entire
backtrace (limited to 10 frames) before printing the first frame, which
I introduced because I was concerned a malformed stack frame was
triggering an out-of-bounds read that stopped me from seeing the
backtrace. I'm not sure this helps, though, since IIRC I didn't even see
the println from before the backtrace call.

This is a breaking API change that's only been implemented for the riscv
half, and worse changes from `u32` to `usize` for no particular reason
beyond that's how I'm used to thinking of these pointers.

But, it's fewer lines of code and I got to feel like I flexed some Rust,
so there's that.

Also changed:

* handles recursive function calls by allowing identical return
  addresses as long as the frame pointer is changing,
* reduces the frame suppression count from 2 to 0, because it was hiding
  actual call frames; maybe this is an `#[inline(never)]` sort of situation?